% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_tables.R
\name{get_analysis_tables}
\alias{get_analysis_tables}
\title{Get preferred analysis tables from a .prism file}
\usage{
get_analysis_tables(prism_file)
}
\arguments{
\item{prism_file}{The path to the prism file}
}
\value{
A named list of data.frames with model parameters
}
\description{
Get preferred analysis tables from a .prism file
}
\examples{
# prism_file is the file path to the .prism file. Analysis data tables in Prism can include multiple (quite often hidden) tabs.
# One of these is usually designated as the preferred tab and this is returned as a list object, commonly a data.frame.
demo_file <- system.file("extdata", "demo_dataset.prism", package = "pRism")
get_analysis_tables(demo_file)
}
