% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_names.R
\name{get_table_names}
\alias{get_table_names}
\title{Get the names of the prism data tables}
\usage{
get_table_names(prism_file)
}
\arguments{
\item{prism_file}{The path to the prism file}
}
\value{
A vector of table names
}
\description{
Get the names of the prism data tables
}
\examples{
# prism_file is the file path to the prism file. It needs to either be a .pzfx or a .prism file.
demo_file <- system.file("extdata", "demo_dataset.prism", package = "pRism")
get_table_names(demo_file)
}
