% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRism-package.R
\docType{package}
\name{pRism-package}
\alias{pRism}
\alias{pRism-package}
\title{pRism: Read Prism Tables Into R}
\description{
This package provides functions to read and parse GraphPad Prism files (.pzfx or .prism) into R data frames, enabling users to analyzse and manipulate their data within the R environment.
}
\author{
\strong{Maintainer}: Miha Kosmac \email{biomiha@gmail.com}

}
\keyword{internal}
