% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRism-package.R
\docType{package}
\name{pRism-package}
\alias{pRism}
\alias{pRism-package}
\title{pRism: Read Prism Tables Into R}
\description{
This package provides functions to read and parse GraphPad Prism files (.pzfx or .prism) into R data frames, enabling users to analyse and manipulate their data within the R environment. It currently consists of two functions - `get_table_names` and `read_prism`, both of which take the file path to a prism file as input and output either the table names or the table contents themselves.
}
\author{
\strong{Maintainer}: Miha Kosmac \email{biomiha@gmail.com}

}
\keyword{internal}
